--- editor/doc/SCsub
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-Import('env')
-
-env.add_source_files(env.editor_sources, "*.cpp")
