+++ editor/doc_data.h
